Conflicts:
	src/java/rexos/mas/equiplet_agent/ProductStepMessage.java
	src/java/rexos/mas/equiplet_agent/behaviours/ScheduleStep.java
	src/java/rexos/mas/hardware_agent/EquipletStepMessage.java
	src/java/rexos/mas/hardware_agent/GripperModule.java
	src/java/rexos/mas/hardware_agent/behaviours/EvaluateDuration.java
	src/java/rexos/mas/service_agent/DummyService.java
	src/java/rexos/mas/service_agent/PickAndPlaceService.java
	src/java/rexos/mas/service_agent/Service.java
	src/java/rexos/mas/service_agent/ServiceAgent.java
	src/java/rexos/mas/service_agent/ServiceStepMessage.java
	src/java/rexos/mas/service_agent/behaviour/CanDoProductStep.java

+++ src/java/rexos/mas/equiplet_agent/NextProductStepTimer.java
@@ -20,7 +20,7 @@ public class NextProductStepTimer extends Timer{
 	 * @var int nextUsedTimeSlot
 	 * The next used time slot.
 	 */
-	private long nextUsedTimeSlot;
+	private int nextUsedTimeSlot;
 
 	/**
 	 * @var int firstTimeSlot
@@ -49,12 +49,12 @@ public class NextProductStepTimer extends Timer{
 	 * Function for setting the next used time slot.
 	 * @param nextUsedTimeSlot the index of the next used timeslot, fill -1 when not used.
 	 */
-	public void setNextUsedTimeSlot(long nextUsedTimeSlot){
+	public void setNextUsedTimeSlot(int nextUsedTimeSlot){
 		this.nextUsedTimeSlot = nextUsedTimeSlot;
 		cancel();
 		if(nextUsedTimeSlot != -1){
-			long startTimeSlot = nextUsedTimeSlot * timeSlotLength + firstTimeSlot;
-			long currentTime = System.currentTimeMillis();
+			int startTimeSlot = nextUsedTimeSlot * timeSlotLength + firstTimeSlot;
+			int currentTime = (int)System.currentTimeMillis();
 			this.schedule(new NextProductStepTask(), startTimeSlot - currentTime);
 		}
 	}
@@ -63,7 +63,7 @@ public class NextProductStepTimer extends Timer{
 	 * Getter for getting the nextUsedTimeSlot.
 	 * @return the next used timeslot
 	 */
-	public long getNextUsedTimeSlot() {
+	public int getNextUsedTimeSlot() {
 		return nextUsedTimeSlot;
 	}
